Reland: [Cast] Consolidate plethora of GYP files into cast.gyp and cast_testing.gypi.
Reland of r265481 (https://codereview.chromium.org/245893005/) with the
following fixes:

1. Re-introduce 'include_tests%' conditional.
2. Conditionally include all Cast utility app targets based on a new GYP
   variable: include_cast_utility_apps
3. Fixed compile errors exposed by landing r265481 on the tree.

Testing:

1. This change was tested by building on Linux:
   `build/gyp_chromium && ninja -C out/Debug All`
2. Confirmed absence of 'include_cast_utility_apps=1' in
   chromium.gyp_env will disable the Cast utility targets.
3. Ran utility apps to confirm the fixes.

R=hclam@chromium.org

Review URL: https://codereview.chromium.org/248083005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265690 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/test/utility/generate_barcode_video.cc
@@ -6,13 +6,14 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include <cstdio>
 #include <cstdlib>
 
+#include "base/logging.h"
 #include "media/base/video_frame.h"
 #include "media/cast/test/utility/barcode.h"
 
 void DumpPlane(scoped_refptr<media::VideoFrame> frame,
                int plane) {
   for (int row = 0; row < frame->rows(plane); row++) {
-    CHECK_EQ(frame->row_bytes(plane),
+    CHECK_EQ(static_cast<size_t>(frame->row_bytes(plane)),
              fwrite(frame->data(plane) + frame->stride(plane) * row,
                     1,
                     frame->row_bytes(plane),
